refactored ImageDescription component. Created EntityCardImage.js and EntityDescription.js components
